refactor(tests): extract qt3tests parser into noir_xpath_inputs package
Move the W3C qt3tests catalogue / test-set XML parser
(parse_test_file, discover_all_test_files, discover_available_functions,
clone_or_update_qt3tests, plus the TestCase dataclass and the QT3_NS
namespace constant) out of scripts/generate_tests.py into a colocated
package scripts/noir_xpath_inputs/ with a dedicated qt3tests submodule.
Mirrors the noir_IEEE754 PR #38 phase-2a layout (commit 807b21d:
`refactor(tests): extract .fptest parser into noir_ieee754_inputs
package`).

Note on Concern 1 of `docs/xpath-input-prep-redesign.md`. The redesign
identifies sissaschool/elementpath's tests/run_w3c_tests.py as the
recommended fork-and-extract source: a usable upstream parser exists
under MIT, and elementpath is already a transitive dep. However, that
file ships only inside the project's git checkout -- it is not packaged
with the installed `elementpath` distribution -- so importing it
directly is not viable without a vendored copy. Per the redesign doc
(sec 6.3) the canonical home of the fork is the workspace package
`tools/noir-test-vectors/qt3tests.py`, scheduled for the cross-repo
extraction round (phase 0 of the cross-repo plan, blocked on D.1).
This commit therefore lifts the in-tree parser as-is into the
per-repo package; the workspace fork-and-extract from
`run_w3c_tests.py` happens once the workspace lift lands. The shape
chosen here (TestCase dataclass + parse_test_file returning
`list[TestCase]`) is compatible with both the elementpath upstream
shape and the cross-format `noir_test_vectors.TestCase` proposed in
the redesign sec 6.3, so no downstream caller will need to change
when the workspace fork lands.

`generate_tests.py` becomes a consumer; the inline `parse_test_file`,
`TestCase`, `discover_all_test_files`, `discover_available_functions`
and `clone_or_update_qt3tests` definitions are deleted and re-imported
from the package. Unused stdlib imports (xml.etree.ElementTree,
dataclasses) are dropped.

Generator output is unchanged: a side-by-side regen against the
post-phase-1 baseline shows only the intrinsic non-determinism that
the historical generator already exhibited (random per-test ID
suffixes inside emitted Noir source); test counts and package list
are identical.
